third_party/WebKit/LayoutTests/css3/device-adapt/opera/constrain-005-expected.txt


FAIL CSS Test: @viewport constrained - min-width greater than max-width. assert_equals: expected 10000 but got 800
